api/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
aniso8601==9.0.1
anyio==3.7.0
APScheduler==3.10.4
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
bcrypt==4.0.1
bidict==0.22.1
CacheControl==0.13.1
cachetools==5.3.1
cairocffi==1.6.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudinary==1.34.0
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.99.1
Fiona==1.9.4.post1
firebase-admin==6.2.0
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
graphql-core==3.2.3
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.5.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.1.0
passlib==1.7.4
pgeocode==0.4.1
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.24.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==1.10.11
PyJWT==2.8.0
pyparsing==3.1.1
pyproj==3.6.0
pyscopg2==66.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
python-socketio==5.8.0
pytz==2023.3.post1
PyYAML==6.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
resend==0.5.2
rsa==4.9
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
starlette-validation-uploadfile==0.2.0
strawberry-graphql==0.204.0
svgwrite==1.4.3
text-unidecode==1.3
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.2
